Uploaded Email Code
Created this program to collect emails of faculty, staff, etc at Illinois Tech. The purpose of creating this was to ensure that the student body hears about this organization and the resources it can provide. By sending the emails to professors and department heads, they can share it with students they are associated with. After using the CSV files generated by this program to send out emails, event attendance increased by over 200%.